+++ Controllers/CorsairPeripheralV2Controller/CorsairPeripheralV2SoftwareController.cpp
@@ -23,7 +23,7 @@ CorsairPeripheralV2SWController::~CorsairPeripheralV2SWController()
 
 void CorsairPeripheralV2SWController::SetLedsDirect(std::vector<RGBColor *>colors)
 {
-    uint16_t count          = colors.size();
+    uint16_t count          = (uint16_t)colors.size();
     uint16_t green          = count;
     uint16_t blue           = count * 2;
     uint16_t length         = count * 3;
